+++ CHANGELOG.md
@@ -7,11 +7,13 @@
 - Add context menu for chat messages to allow rerolls for valid d20 rolls from the chat pane.
 - Add CHANGELOG.md.
 - Update Github Workflows.
+- Add message handler.
 
 ### Fixed
 
 - Fix SASS deprecation warnings.
+- Fix Tidy 5e rendering.
 
 ### Security
 
-- Fix NPM package vulnerabilities
+- Fix NPM package vulnerabilities.